feat: remove no required code
