+++ src/handlers/dkg_round_1.rs
@@ -22,7 +22,7 @@ use ironfish_frost::dkg;
 use ironfish_frost::participant::{Identity, Secret};
 use ledger_device_sdk::io::{Comm, Event};
 use crate::handlers::dkg_get_identity::compute_dkg_secret;
-use crate::handlers::sign_tx::TxContext;
+use crate::contex::TxContext;
 
 const MAX_TRANSACTION_LEN: usize = 2040;
 const MAX_APDU_SIZE: usize = 253;
